Refactor PixelsBuilder into its own module
